requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain_google_genai
flask_cors
transformers
torch
numpy
accelerate
huggingface_hub
streamlit
-e .